+++ Projeto-01_Cifra-de-Vigenere/ciphertext.txt
@@ -1,1 +1,1 @@
-hokpu, ze soj vwwu wds xdj tlpa wcri ksh bfcwd, quka crlsid cokp. vv fwav rj wzrp, whwrzyj vzd houpq vfcva sj pvv enwuwh. ze soj owhv dbhvcqkce djr eka izsafj', groghsid' oeo poeyhng' vpocwv kvch occhwrp fhcjpg, ksh gkchah pplhp. eh hdo yzw ple pvv pwb sdz r ehotv ycre pvizzj fghn ytv gyzxhrvcv. yp zfvh whkpqpwfy pc sligvwi.
+mhbv yuvv fuegs, qlnuuz egy degyyh tizha prlc xb balgh, xb dalgdc xbv oiegcqcerg. pkssih o jr, oiagilcagc mrx ouidza chzmm uspynv. dn sdu wodxr; mm eodlnfoe qrg wburayvuce r gqfrlfam hgfuida zrfvmxnv q n fgumlbtu, rqmruhm. iolcn qmfbu, gnv a uraqg fodlrjohu xam chgmxn qmjn ddygd eious if cyverg. pkoyuid m nwszwnr.
