+++ Projeto-01_Cifra-de-Vigenere/plaintexts backup/pt_plaintext.txt
